add abstract mapper and class CommentMapper

+++ src/main/java/org/example/newsmanager/service/impl/CommentServiceImpl.java
@@ -1,26 +1,27 @@
 package org.example.newsmanager.service.impl;
 
 import lombok.RequiredArgsConstructor;
-import org.example.newsmanager.models.bean.CommentDataForEditBean;
-import org.example.newsmanager.models.bean.CommentInfoBean;
+import org.example.newsmanager.models.bean.CommentResponse;
 import org.example.newsmanager.models.entity.CommentEntity;
 import org.example.newsmanager.repository.CommentDAO;
 import org.example.newsmanager.repository.exception.DAOException;
 import org.example.newsmanager.service.CommentService;
 import org.example.newsmanager.service.exception.ServiceException;
 import org.example.newsmanager.service.util.converter.DateConverter;
+import org.example.newsmanager.service.util.mapper.Mapper;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class CommentServiceImpl implements CommentService {
 	private final CommentDAO commentDAO;
-	private final DateConverter dateConverter;
+	private final Mapper <CommentEntity, CommentResponse> mapper;
 
 	@Override
 	@Transactional
@@ -39,22 +40,14 @@ public class CommentServiceImpl implements CommentService {
 
 	@Override
 	@Transactional(readOnly = true)
-	public List<CommentInfoBean> getCommentsFromNewsById(String newsId, Locale locale) throws ServiceException {
+	public List<CommentResponse> getCommentsFromNewsById(String newsId, Locale locale) throws ServiceException {
 		try {
 			int id = Integer.parseInt(newsId);
 			List<CommentEntity> commentsEntity = commentDAO.getCommentsFromNewsById(id);
-			List<CommentInfoBean> comments = new ArrayList<>();
-			
-			for(CommentEntity comment: commentsEntity) {
-				CommentInfoBean commentInfo = new CommentInfoBean();
-				commentInfo.setId(comment.getId());
-				commentInfo.setText(comment.getText());
-				commentInfo.setUsername(comment.getUserEntity().getLogin());
-				String formatPublicationDate = dateConverter.getFormatDateByComment(comment.getPublicationDate(), locale);
-				commentInfo.setPublicationDate(formatPublicationDate);
-				comments.add(commentInfo);
-			}
-			return comments;
+
+			return commentsEntity.stream()
+					.map(commentEntity -> mapper.mapToDto(commentEntity, locale))
+					.toList();
 		}
 		catch(DAOException | IllegalArgumentException e) {
 			throw new ServiceException(e);
@@ -66,6 +59,12 @@ public class CommentServiceImpl implements CommentService {
 	public void deleteCommentById(String commentId) throws ServiceException {
 		try {
 			int id = Integer.parseInt(commentId);
+
+			Optional<CommentEntity> optionalComment = commentDAO.getCommentById(id);
+			if (optionalComment.isEmpty()){
+				throw new ServiceException("Comment with id `" + commentId + "` not found.");
+			}
+
 			commentDAO.deleteCommentById(id);
 		}
 		catch(IllegalArgumentException | DAOException e) {
@@ -75,14 +74,16 @@ public class CommentServiceImpl implements CommentService {
 	
 	@Override
 	@Transactional(readOnly = true)
-	public CommentDataForEditBean getCommentById(String commentId) throws ServiceException{
+	public CommentResponse getCommentById(String commentId) throws ServiceException{
 		try {
 			int id = Integer.parseInt(commentId);
-			CommentEntity commentEntity = commentDAO.getCommentById(id);
-			CommentDataForEditBean comment = new CommentDataForEditBean();
-			comment.setId(commentEntity.getId());
-			comment.setText(commentEntity.getText());
-			return comment;
+			Optional<CommentEntity> optionalCommentEntity = commentDAO.getCommentById(id);
+
+			if (optionalCommentEntity.isEmpty()){
+				throw new ServiceException("Comment with id `" + commentId + "` not found.");
+			}
+
+			return mapper.mapToDto(optionalCommentEntity.get(), null);
 		}
 		catch(DAOException e) {
 			throw new ServiceException(e);
@@ -94,6 +95,12 @@ public class CommentServiceImpl implements CommentService {
 	public void editCommentById(String commentId, String text) throws ServiceException {
 		try{
 			int id = Integer.parseInt(commentId);
+
+			Optional<CommentEntity> optionalComment = commentDAO.getCommentById(id);
+			if (optionalComment.isEmpty()){
+				throw new ServiceException("Comment with id `" + commentId + "` not found.");
+			}
+
 			commentDAO.editCommentById(id, text);
 		}
 		catch (DAOException e){
